renames depchekc file
